# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API GIMA",
        "description": "Documentación de la API para el proyecto GIMA (Gestión de Inventario y Mantenimiento)",
        "contact": {
            "email": "admin@gima.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor API Principal"
        }
    ],
    "paths": {
...
        "/api/catalogo/articulos": {
            "post": {
                "tags": [
                    "Catálogo - Artículos"
                ],
                "summary": "Crear artículo",
                "operationId": "65016e90edbd487dc06816eac8f93d34",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tipo": {
                                        "type": "string"
                                    },
                                    "marca": {
                                        "type": "string"
                                    },
                                    "modelo": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Artículo creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Articulo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/catalogo/articulos/{id}": {
            "get": {
                "tags": [
                    "Catálogo - Artículos"
                ],
                "summary": "Ver artículo",
                "operationId": "8bbaf219b247018034f4bc1a2b21f7d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artículo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Articulo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Catálogo - Artículos"
                ],
                "summary": "Actualizar artículo",
                "operationId": "1ebcccc112b188622fa03279d27d819b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tipo": {
                                        "type": "string"
                                    },
                                    "marca": {
                                        "type": "string"
                                    },
                                    "modelo": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Artículo actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Articulo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Catálogo - Artículos"
                ],
                "summary": "Eliminar artículo",
                "operationId": "97b8d1b6815c7ee5b8016ff27774d844",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado"
                    },
                    "409": {
                        "description": "Conflicto al eliminar"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/general/notificaciones": {
            "get": {
                "tags": [
                    "General - Notificaciones"
                ],
                "summary": "Listar notificaciones",
                "operationId": "15c8d39f33074452f2ac3395db90a407",
                "responses": {
                    "200": {
                        "description": "Lista de notificaciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Notificacion"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "General - Notificaciones"
                ],
                "summary": "Crear notificación",
                "operationId": "1c5c34f29c7fc3862d76db9e2c0699db",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "usuario_id": {
                                        "type": "integer"
                                    },
                                    "contenido": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notificación creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notificacion"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/general/notificaciones/{id}": {
            "get": {
                "tags": [
                    "General - Notificaciones"
                ],
                "summary": "Ver notificación",
                "operationId": "3a521388bda407f2bdc059fa274a90f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificación encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notificacion"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Articulo": {
                "title": "Articulo",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "tipo": {
                        "type": "string"
                    },
                    "marca": {
                        "type": "string"
                    },
                    "modelo": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Notificacion": {
                "title": "Notificacion",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "usuario_id": {
                        "type": "integer"
                    },
                    "contenido": {
                        "type": "string"
                    },
                    "leido": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}